+++ CHANGES.md
@@ -14,6 +14,7 @@ Change Log
 * Fixed a bug where WMS items caused type errors in newer babel and typescript builds, due to mixed mixin methods on DiffableMixin & DiscretelyTimeVaryingMixin
 * Fixed a bug where KmlCatalogItem did not use the proxy for any urls.
 * Add support for `CkanCatalogGroup` and `CkanItemReference`.
+* Added unit test to ensure getAncestors behaviour
 * (💫The next rad feature💫 but please be mostly bug fixes from now until June!)
 
 #### mobx-31